+++ python/rateslib/dual/quadratic.py
@@ -1,7 +1,11 @@
-from typing import Any
+from __future__ import annotations
 
-from rateslib.dual import DualTypes
-from rateslib.solver.utils import _solver_result
+from typing import TYPE_CHECKING, Any
+
+from rateslib.dual.newton import _solver_result
+
+if TYPE_CHECKING:
+    from rateslib.dual.utils import DualTypes
 
 # Licence: Creative Commons - Attribution-NonCommercial-NoDerivatives 4.0 International
 # Commercial use of this code, and/or copying and redistribution is prohibited.
