+++ python/tests/test_dualpy.py
@@ -7,7 +7,6 @@ from packaging import version
 from rateslib.dual import (
     Dual,
     Dual2,
-    _abs_float,
     dual_exp,
     dual_inv_norm_cdf,
     dual_log,
@@ -16,6 +15,7 @@ from rateslib.dual import (
     gradient,
     set_order,
 )
+from rateslib.dual.utils import _abs_float
 
 
 @pytest.fixture
